requirements.txt
appdirs==1.4.4
astroid==2.4.2
attrs==20.3.0
autopep8==1.5.4
cached-property==1.5.2
certifi==2020.12.5
chardet==4.0.0
defusedxml==0.6.0
idna==2.10
isodate==0.6.0
isort==5.7.0
lazy-object-proxy==1.4.3
lxml==4.6.2
mccabe==0.6.1
pycodestyle==2.6.0
pylint==2.6.0
python-dotenv==0.15.0
pytz==2020.5
requests==2.25.1
requests-file==1.5.1
requests-toolbelt==0.9.1
six==1.15.0
toml==0.10.2
urllib3==1.26.2
wrapt==1.12.1
zeep==4.0.0
